requirements.txt
### External Python libraries to install in venv
# TODO: add versions

# dev stuff
ipykernel
black
black[jupyter]

# blockchain stuff
eth-brownie
web3

# server stuff
flask
flask_cors
gunicorn
python-dotenv

# db
sqlalchemy
psycopg2
